% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbDatatoREDCap.R
\name{initConstants}
\alias{initConstants}
\title{Initialize Configuration Constants from TOML Files}
\usage{
initConstants()
}
\description{
This function loads configuration settings from two TOML files: one for the module
configuration and one for the database configuration.
It sets all variables defined in these files in the global context. The function is
designed to be run at the start of a session to set up necessary configurations.
The path to the module configuration file is determined based on whether the session
is interactive or not, defaulting to './R-db2frontend/db2frontend_config.toml' in
interactive sessions. The path to the database configuration file is specified by
the global variable PATH_TO_DB_CONFIG_TOML.
}
\seealso{
\code{\link[etlutils]{initConstants}} for details on how configuration settings are loaded and set.
}
